Dungeon Game
Created a dungeon game with a 2-dimensional map. Placed the player, a
door, and a monster into random spots in the map. The player can move
around, but the game ends when he/she either finds the door or gets
eaten by the Grue (monster). The door is the win condition.